+++ test/integration/default/serverspec/bower_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 describe 'bower' do
-  describe command('/usr/local/bin/bower -v') do
+  describe command('/bin/bower -v') do
     its(:exit_status) { should eq 0 }
     its(:stdout) { should match /^[\w\.]+$/ }
   end