+++ test/integration/default/serverspec/essentials_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 describe 'essentials' do
-  describe yumrepo('epel') do
-    it { should exist }
+  describe command("yum repolist all -C | grep '^!\\?epel'") do
+    its(:exit_status) { should eq 0 }
   end
 end